+++ lib/bindings.dart
@@ -1,12 +1,9 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-/// Provides bindings to NodeJS APIs.
+/// Provides bindings to different NodeJS APIs.
 ///
-/// Some APIs, like 'fs' and 'http' must be `require`d to use:
-///
-///     FS nodeFS = require('fs');
-///     HTTP nodeHTTP = require('http');
+@Deprecated('Import node_interop/node_interop.dart directly.')
 library node_interop.bindings;
 
 export 'src/bindings/events.dart';